Move inline Rust tests into sibling files
